fix(review-loop): drain agent log writes in dispose (CI ENOENT flake)
onLine fired `void appendFile(logPath, ...)` per agent output line and forgot
it. When the caller's temp dir was removed while the write was still queued
(test afterEach cleanup, run teardown), the rejection surfaced as an unhandled
ENOENT attributed to whatever code was running when it landed — the CI flake
failing 'concurrent runAgent calls ... do not race' and 'timeout throw carries
accumulated usage' (always the test right after an onLine-emitting one).

Log appends now serialize through a best-effort promise chain (enqueueLog),
dispose() drains it, and runAgent's finally awaits dispose(), so no log write
outlives the run. The stderr log write goes through the same chain to preserve
ordering. Line-handler unit extracted to review-loop/src/line-handler.ts
(max-lines); AgentUsage/LineSink stay re-exported from agent-runner.js.
